+++ molo/profiles/tests/test_views.py
@@ -4,12 +4,17 @@ from datetime import date
 from django.conf.urls import patterns, url, include
 from django.conf import settings
 from django.contrib.auth.models import User
+from django.contrib.auth.tokens import default_token_generator
 from django.core.urlresolvers import reverse
+from django.http import QueryDict
 from django.test import TestCase, override_settings, Client
 
 from molo.profiles.forms import (
-    RegistrationForm, EditProfileForm, ProfilePasswordChangeForm)
-from molo.profiles.models import UserProfile
+    RegistrationForm, EditProfileForm,
+    ProfilePasswordChangeForm, ForgotPasswordForm)
+from molo.profiles.models import (
+    SecurityQuestion, SecurityAnswer, UserProfile)
+from molo.core.models import PageTranslation, SiteLanguage
 from molo.core.tests.base import MoloTestCaseMixin
 
 from wagtail.wagtailcore.models import Site
@@ -32,6 +37,8 @@ class RegistrationViewTest(TestCase, MoloTestCaseMixin):
     def setUp(self):
         self.client = Client()
         self.mk_main()
+        # Creates Main language
+        SiteLanguage.objects.create(locale='en')
 
     def test_register_view(self):
         response = self.client.get(reverse('molo.profiles:user_register'))
@@ -335,6 +342,38 @@ class RegistrationViewTest(TestCase, MoloTestCaseMixin):
 
         self.assertContains(response, expected_validation_message)
 
+    def test_security_questions(self):
+        site = Site.objects.get(is_default_site=True)
+        settings = SettingsProxy(site)
+
+        SecurityQuestion.objects.create(
+            title="What is your name?",
+            slug="what-is-your-name",
+            path="0002",
+            depth=1,
+        )
+
+        profile_settings = settings['profiles']['UserProfilesSettings']
+        profile_settings.show_security_question_fields = True
+        profile_settings.security_questions_required = True
+        profile_settings.save()
+
+        response = self.client.get(reverse('molo.profiles:user_register'))
+        self.assertContains(response, "What is your name")
+
+        # register with security questions
+        response = self.client.post(
+            reverse("molo.profiles:user_register"),
+            {
+                "username": "tester",
+                "password": "0000",
+                "question_0": "answer",
+                'terms_and_conditions': True
+            },
+            follow=True
+        )
+        self.assertEqual(response.status_code, 200)
+
 
 @override_settings(
     ROOT_URLCONF='molo.profiles.tests.test_views')
@@ -491,9 +530,10 @@ class ProfileDateOfBirthEditTest(MoloTestCaseMixin, TestCase):
 
 @override_settings(
     ROOT_URLCONF='molo.profiles.tests.test_views')
-class ProfilePasswordChangeViewTest(TestCase):
+class ProfilePasswordChangeViewTest(TestCase, MoloTestCaseMixin):
 
     def setUp(self):
+        self.mk_main()
         self.user = User.objects.create_user(
             username='tester',
             email='tester@example.com',
@@ -540,3 +580,277 @@ class ProfilePasswordChangeViewTest(TestCase):
         # Avoid cache by loading from db
         user = User.objects.get(pk=self.user.pk)
         self.assertTrue(user.check_password('1234'))
+
+
+@override_settings(
+    ROOT_URLCONF="molo.profiles.tests.test_views",
+)
+class ForgotPasswordViewTest(TestCase, MoloTestCaseMixin):
+
+    def setUp(self):
+        self.mk_main()
+        self.client = Client()
+        self.user = User.objects.create_user(
+            username="tester",
+            email="tester@example.com",
+            password="0000")
+        # Creates Main language
+        SiteLanguage.objects.create(locale='en')
+
+        # create a few security questions
+        q1 = SecurityQuestion.objects.create(
+            title="How old are you?",
+            slug="how-old-are-you",
+            path="0002",
+            depth=1,
+        )
+
+        # create answers for this user
+        self.a1 = SecurityAnswer.objects.create(
+            user=self.user.profile, question=q1, answer="20"
+        )
+
+    def test_view(self):
+        response = self.client.get(
+            reverse("molo.profiles:forgot_password"))
+        form = response.context["form"]
+        self.assertTrue(isinstance(form, ForgotPasswordForm))
+
+    def test_unidentified_user_gets_error(self):
+        error_message = "The username and security question(s) combination " \
+                        "do not match."
+        response = self.client.post(
+            reverse("molo.profiles:forgot_password"), {
+                "username": "bogus",
+                "question_0": "20",
+            }
+        )
+        self.failUnless(error_message in response.content)
+
+    def test_suspended_user_gets_error(self):
+        error_message = "The username and security question(s) combination " \
+                        "do not match."
+        self.user.is_active = False
+        self.user.save()
+        response = self.client.post(
+            reverse("molo.profiles:forgot_password"), {
+                "username": "tester",
+                "question_0": "20",
+            }
+        )
+        self.failUnless(error_message in response.content)
+        self.user.is_active = True
+        self.user.save()
+
+    def test_incorrect_security_answer_gets_error(self):
+        error_message = "The username and security question(s) combination " \
+                        "do not match."
+        response = self.client.post(
+            reverse("molo.profiles:forgot_password"), {
+                "username": "tester",
+                "question_0": "21",
+            }
+        )
+        self.failUnless(error_message in response.content)
+
+    def test_too_many_retries_result_in_error(self):
+        error_message = "Too many attempts"
+        site = Site.objects.get(is_default_site=True)
+        settings = SettingsProxy(site)
+        profile_settings = settings['profiles']['UserProfilesSettings']
+
+        # post more times than the set number of retries
+        for i in range(profile_settings.password_recovery_retries + 5):
+            response = self.client.post(
+                reverse("molo.profiles:forgot_password"), {
+                    "username": "bogus",
+                    "question_0": "20",
+                }
+            )
+
+        self.failUnless(error_message in response.content)
+
+    def test_correct_username_and_answer_results_in_redirect(self):
+        response = self.client.post(
+            reverse("molo.profiles:forgot_password"), {
+                "username": "tester",
+                "question_0": "20",
+            },
+            follow=True
+        )
+        self.assertContains(response, "Reset PIN")
+
+
+class TranslatedSecurityQuestionsTest(TestCase, MoloTestCaseMixin):
+
+    def setUp(self):
+        self.mk_main()
+        self.client = Client()
+        self.questions_section = self.mk_section(
+            self.section_index, title='Security Questions')
+
+        # Creates Main language
+        SiteLanguage.objects.create(locale="en")
+
+        # Creates translation Language
+        self.french = SiteLanguage.objects.create(locale="fr")
+
+        # create a few security questions
+        self.q1 = SecurityQuestion.objects.create(
+            title="How old are you?",
+            slug="how-old-are-you",
+            path="0002",
+            depth=1,
+        )
+
+    def test_translated_question_appears_on_registration(self):
+        # make translation for the security question
+        fr_question = SecurityQuestion.objects.create(
+            title="How old are you in french",
+            slug="how-old-are-you-in-french",
+            path="0003",
+            depth=1,
+        )
+        language_relation = fr_question.languages.first()
+        language_relation.language = self.french
+        language_relation.save()
+        fr_question.save_revision().publish()
+        PageTranslation.objects.get_or_create(
+            page=self.q1, translated_page=fr_question)
+
+        self.client.get('/locale/fr/')
+        response = self.client.get(reverse("molo.profiles:forgot_password"))
+        self.assertContains(response, "How old are you in french")
+
+        # switch locale to english and check that the english question
+        # is asked
+        self.client.get('/locale/en/')
+        response = self.client.get(reverse("molo.profiles:forgot_password"))
+        self.failIf("How old are you in french" in response.content)
+
+
+class ResetPasswordViewTest(TestCase, MoloTestCaseMixin):
+    def setUp(self):
+        self.mk_main()
+        self.client = Client()
+        self.user = User.objects.create_user(
+            username="tester",
+            email="tester@example.com",
+            password="0000")
+
+        # Creates Main language
+        SiteLanguage.objects.create(locale='en')
+
+        # create a few security questions
+        q1 = SecurityQuestion.objects.create(
+            title="How old are you?",
+            slug="how-old-are-you",
+            path="0002",
+            depth=1,
+        )
+
+        # create answers for this user
+        self.a1 = SecurityAnswer.objects.create(
+            user=self.user.profile, question=q1, answer="20"
+        )
+
+    def proceed_to_reset_password_page(self):
+        expected_token = default_token_generator.make_token(self.user)
+        expected_query_params = QueryDict(mutable=True)
+        expected_query_params["user"] = self.user.username
+        expected_query_params["token"] = expected_token
+        expected_redirect_url = "{0}?{1}".format(
+            reverse("molo.profiles:reset_password"),
+            expected_query_params.urlencode()
+        )
+
+        response = self.client.post(
+            reverse("molo.profiles:forgot_password"), {
+                "username": "tester",
+                "question_0": "20",
+            }
+        )
+
+        self.assertRedirects(response, expected_redirect_url)
+
+        return expected_token, expected_redirect_url
+
+    def test_reset_password_view_pin_mismatch(self):
+        expected_token, expected_redirect_url = \
+            self.proceed_to_reset_password_page()
+
+        response = self.client.post(expected_redirect_url, {
+            "username": self.user.username,
+            "token": expected_token,
+            "password": "1234",
+            "confirm_password": "4321"
+        })
+
+        self.assertContains(response, "The two PINs that you entered do not "
+                                      "match. Please try again.")
+
+    def test_reset_password_view_requires_query_params(self):
+        response = self.client.get(reverse("molo.profiles:reset_password"))
+        self.assertEqual(403, response.status_code)
+
+    def test_reset_password_view_invalid_username(self):
+        expected_token, expected_redirect_url = \
+            self.proceed_to_reset_password_page()
+
+        response = self.client.post(expected_redirect_url, {
+            "username": "invalid",
+            "token": expected_token,
+            "password": "1234",
+            "confirm_password": "1234"
+        })
+
+        self.assertEqual(403, response.status_code)
+
+    def test_reset_password_view_inactive_user(self):
+        expected_token, expected_redirect_url = \
+            self.proceed_to_reset_password_page()
+
+        self.user.is_active = False
+        self.user.save()
+
+        response = self.client.post(expected_redirect_url, {
+            "username": self.user.username,
+            "token": expected_token,
+            "password": "1234",
+            "confirm_password": "1234"
+        })
+
+        self.assertEqual(403, response.status_code)
+
+    def test_reset_password_view_invalid_token(self):
+        expected_token, expected_redirect_url = \
+            self.proceed_to_reset_password_page()
+
+        response = self.client.post(expected_redirect_url, {
+            "username": self.user.username,
+            "token": "invalid",
+            "password": "1234",
+            "confirm_password": "1234"
+        })
+
+        self.assertEqual(403, response.status_code)
+
+    def test_happy_path(self):
+        expected_token, expected_redirect_url = \
+            self.proceed_to_reset_password_page()
+
+        response = self.client.post(expected_redirect_url, {
+            "username": self.user.username,
+            "token": expected_token,
+            "password": "1234",
+            "confirm_password": "1234"
+        })
+
+        self.assertRedirects(
+            response,
+            reverse("molo.profiles:reset_password_success")
+        )
+
+        self.assertTrue(
+            self.client.login(username="tester", password="1234")
+        )